fix: promise all
